Add lookup of users by email address
I added a new macro to do quick atomic redis sets the act based
on whether the set was successful or now.  Now, creating a user
checks the username atomically then checks the email address
atomically.  If the email check fails, we delete the user key
and report failure.

All 13* tests pass.

* There are more than 13 tests, but eunit only reports suite tests
when running in this mode, not each comparison as when running
tests in-line directly.